requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
audio-recorder-streamlit==0.0.8
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.10.0
decorator==5.1.1
easyocr==1.7.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
googletrans==3.0.0
gTTS==2.3.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.31.1
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1
numpy==1.25.0
openai==0.27.8
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pyclipper==1.3.0.post4
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzbar==0.1.9
requests==2.31.0
requests-toolbelt==1.0.0
requires.io==0.2.6
rfc3986==1.5.0
rich==13.4.2
roboflow==1.0.9
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SpeechRecognition==3.10.0
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
ultralytics==8.0.118
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wget==3.2
yarl==1.9.2
zipp==3.15.0
